Move DeploymentsArrays out
